requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
filelock==3.9.0
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
mypy==0.991
mypy-extensions==0.4.3
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
platformdirs==2.6.2
pycodestyle==2.10.0
pydocstyle==6.3.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
requests==2.28.2
six==1.16.0
snowballstemmer==2.2.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
virtualenv==20.17.1
Werkzeug==2.2.2
